+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/OpenableSourceInfo.java
@@ -207,8 +207,9 @@ public class OpenableSourceInfo {
                         || !(
                             extension.equals(".swf")
                             || extension.equals(".spl")
+                            || extension.equals(".swt")
                             || extension.equals(".gfx") 
-                            || extension.equals(".abc")
+                            || extension.equals(".abc")                            
                             )
                     );
         }