+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/DefineVideoTag.java
@@ -19,66 +19,50 @@ package com.jpexs.decompiler.flash.tags;
 import com.jpexs.decompiler.flash.SWF;
 import com.jpexs.decompiler.flash.SWFInputStream;
 import com.jpexs.decompiler.flash.SWFOutputStream;
-import com.jpexs.decompiler.flash.tags.base.CharacterIdTag;
 import com.jpexs.decompiler.flash.types.BasicType;
 import com.jpexs.decompiler.flash.types.annotations.SWFType;
 import com.jpexs.decompiler.flash.types.annotations.SWFVersion;
 import com.jpexs.helpers.ByteArrayRange;
 import java.io.IOException;
-import java.util.Set;
 
 /**
- * NameCharacter tag - Assign a Library name to a character.
- * Used in Flash Templates.
+ * DefineVideo tag - references QuickTime video by URL.
  *
  * @author JPEXS
  */
 @SWFVersion(from = 3)
-public class NameCharacterTag extends Tag implements CharacterIdTag {
+public class DefineVideoTag extends Tag {
 
-    public static final int ID = 40;
+    public static final int ID = 38;
 
-    public static final String NAME = "NameCharacter";
-
+    public static final String NAME = "DefineVideo";   
     
-    public static final int TYPE_BITMAP = 1;
-    public static final int TYPE_SYMBOL = 6;
-    public static final int TYPE_SOUND = 0xFFFF;
-    
-    
-    /**
-     * ID of character to name
-     */
     @SWFType(BasicType.UI16)
-    public int characterId = 0;
-
-    /**
-     * Name of the character
-     */
-    public String name = "Symbol";
-
-    @SWFType(BasicType.UI16)    
-    public int type = TYPE_SYMBOL;
+    public int videoId;
     
+    /**
+     * QuickTime video path (.mov)
+     */
+    public String videoFileName;
+
     /**
      * Constructor
      *
      * @param swf SWF
      */
-    public NameCharacterTag(SWF swf) {
+    public DefineVideoTag(SWF swf) {
         super(swf, ID, NAME, null);
     }
 
-    public NameCharacterTag(SWFInputStream sis, ByteArrayRange data) throws IOException {
+    public DefineVideoTag(SWFInputStream sis, ByteArrayRange data) throws IOException {
         super(sis.getSwf(), ID, NAME, data);
         readData(sis, data, 0, false, false, false);
     }
 
     @Override
     public final void readData(SWFInputStream sis, ByteArrayRange data, int level, boolean parallel, boolean skipUnusualTags, boolean lazy) throws IOException {
-        characterId = sis.readUI16("characterId");
-        name = sis.readString("name");
-        type = sis.readUI16("type");
+        videoId = sis.readUI16("videoId");
+        videoFileName = sis.readString("videoFileName");
     }
 
     /**
@@ -89,23 +73,7 @@ public class NameCharacterTag extends Tag implements CharacterIdTag {
      */
     @Override
     public void getData(SWFOutputStream sos) throws IOException {
-        sos.writeUI16(characterId);
-        sos.writeString(name);
-        sos.writeUI16(type);
-    }
-
-    @Override
-    public int getCharacterId() {
-        return characterId;
-    }
-
-    @Override
-    public void setCharacterId(int characterId) {
-        this.characterId = characterId;
-    }
-
-    @Override
-    public void getNeededCharacters(Set<Integer> needed, SWF swf) {
-        needed.add(characterId);
+        sos.writeUI16(videoId);
+        sos.writeString(videoFileName);
     }
 }