+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/ShowFrameTag.java
@@ -52,6 +52,7 @@ public class ShowFrameTag extends Tag {
             add(VideoFrameTag.ID);
             add(SoundStreamBlockTag.ID);
             add(SetTabIndexTag.ID);
+            add(GenCommandTag.ID);
             /*add(SoundStreamHeadTag.ID);
              add(SoundStreamHead2Tag.ID);*/
         }