+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/Tag.java
@@ -292,7 +292,7 @@ public abstract class Tag implements NeedsCharacters, Exportable, Serializable {
 
     private static volatile Integer[] knownTagIds;
 
-    private static volatile Map<Integer, TagTypeInfo> knownTagInfosById;
+    private static volatile Map<Integer, List<TagTypeInfo>> knownTagInfosById;
 
     private static volatile Map<String, TagTypeInfo> knownTagInfosByName;
 
@@ -319,11 +319,11 @@ public abstract class Tag implements NeedsCharacters, Exportable, Serializable {
      * Gets known classes.
      * @return Known classes
      */
-    public static Map<Integer, TagTypeInfo> getKnownClasses() {
+    public static Map<Integer, List<TagTypeInfo>> getKnownClasses() {
         if (knownTagInfosById == null) {
             synchronized (lockObject) {
                 if (knownTagInfosById == null) {
-                    Map<Integer, TagTypeInfo> map = new HashMap<>();
+                    Map<Integer, List<TagTypeInfo>> map = new HashMap<>();
                     Map<String, TagTypeInfo> map2 = new HashMap<>();
                     addTagInfo(map, map2, CSMSettingsTag.ID, CSMSettingsTag.class, CSMSettingsTag.NAME);
                     addTagInfo(map, map2, DebugIDTag.ID, DebugIDTag.class, DebugIDTag.NAME);
@@ -405,6 +405,14 @@ public abstract class Tag implements NeedsCharacters, Exportable, Serializable {
                     addTagInfo(map, map2, DefineSubImage.ID, DefineSubImage.class, DefineSubImage.NAME);
                     addTagInfo(map, map2, ExporterInfo.ID, ExporterInfo.class, ExporterInfo.NAME);
                     addTagInfo(map, map2, FontTextureInfo.ID, FontTextureInfo.class, FontTextureInfo.NAME);
+                    
+                    addTagInfo(map, map2, DefineVideoTag.ID, DefineVideoTag.class, DefineVideoTag.NAME);
+                    addTagInfo(map, map2, GenCommandTag.ID, GenCommandTag.class, GenCommandTag.NAME);
+                    addTagInfo(map, map2, FontRefTag.ID, FontRefTag.class, FontRefTag.NAME);
+                    addTagInfo(map, map2, DefineTextFormatTag.ID, DefineTextFormatTag.class, DefineTextFormatTag.NAME);
+                    addTagInfo(map, map2, NameCharacterTag.ID, NameCharacterTag.class, NameCharacterTag.NAME);
+                    addTagInfo(map, map2, CharacterSetTag.ID, CharacterSetTag.class, CharacterSetTag.NAME);
+                    addTagInfo(map, map2, SerialNumberTag.ID, SerialNumberTag.class, SerialNumberTag.NAME);
                     knownTagInfosById = map;
                     knownTagInfosByName = map2;
                 }
@@ -426,8 +434,11 @@ public abstract class Tag implements NeedsCharacters, Exportable, Serializable {
         return knownTagInfosByName;
     }
 
-    private static void addTagInfo(Map<Integer, TagTypeInfo> map, Map<String, TagTypeInfo> map2, int id, Class cls, String name) {
-        map.put(id, new TagTypeInfo(id, cls, name));
+    private static void addTagInfo(Map<Integer, List<TagTypeInfo>> map, Map<String, TagTypeInfo> map2, int id, Class cls, String name) {
+        if (!map.containsKey(id)) {
+            map.put(id, new ArrayList<>());
+        }
+        map.get(id).add(new TagTypeInfo(id, cls, name));
         map2.put(name, new TagTypeInfo(id, cls, name));
     }