+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/text/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Text tags.
+ * Text tag types.
  */
 package com.jpexs.decompiler.flash.tags.text;