+++ src/com/jpexs/decompiler/flash/gui/tagtree/AbstractTagTree.java
@@ -53,6 +53,8 @@ import com.jpexs.decompiler.flash.tags.DefineScalingGridTag;
 import com.jpexs.decompiler.flash.tags.DefineSceneAndFrameLabelDataTag;
 import com.jpexs.decompiler.flash.tags.DefineSoundTag;
 import com.jpexs.decompiler.flash.tags.DefineSpriteTag;
+import com.jpexs.decompiler.flash.tags.DefineTextFormatTag;
+import com.jpexs.decompiler.flash.tags.DefineTextTag;
 import com.jpexs.decompiler.flash.tags.DefineVideoStreamTag;
 import com.jpexs.decompiler.flash.tags.DoActionTag;
 import com.jpexs.decompiler.flash.tags.DoInitActionTag;
@@ -67,6 +69,7 @@ import com.jpexs.decompiler.flash.tags.ImportAssets2Tag;
 import com.jpexs.decompiler.flash.tags.ImportAssetsTag;
 import com.jpexs.decompiler.flash.tags.JPEGTablesTag;
 import com.jpexs.decompiler.flash.tags.MetadataTag;
+import com.jpexs.decompiler.flash.tags.NameCharacterTag;
 import com.jpexs.decompiler.flash.tags.PlaceImagePrivateTag;
 import com.jpexs.decompiler.flash.tags.PlaceObject2Tag;
 import com.jpexs.decompiler.flash.tags.PlaceObject3Tag;
@@ -101,6 +104,7 @@ import com.jpexs.decompiler.flash.tags.base.PlaceObjectTypeTag;
 import com.jpexs.decompiler.flash.tags.base.RemoveTag;
 import com.jpexs.decompiler.flash.tags.base.ShapeTag;
 import com.jpexs.decompiler.flash.tags.base.SoundStreamHeadTypeTag;
+import com.jpexs.decompiler.flash.tags.base.StaticTextTag;
 import com.jpexs.decompiler.flash.tags.base.SymbolClassTypeTag;
 import com.jpexs.decompiler.flash.tags.base.TextTag;
 import com.jpexs.decompiler.flash.tags.gfx.DefineCompactedFont;
@@ -1111,10 +1115,13 @@ public abstract class AbstractTagTree extends JTree {
 
     public static List<Integer> getMappedTagIdsForClass(Class<?> cls) {
         if (cls == DefineSpriteTag.class) {
-            return Arrays.asList(DefineScalingGridTag.ID, DoInitActionTag.ID);
+            return Arrays.asList(DefineScalingGridTag.ID, DoInitActionTag.ID, NameCharacterTag.ID);
         }
         if (FontTag.class.isAssignableFrom(cls)) {
             return Arrays.asList(DefineFontNameTag.ID, DefineFontAlignZonesTag.ID, DefineFontInfoTag.ID, DefineFontInfo2Tag.ID);
+        }
+        if (StaticTextTag.class.isAssignableFrom(cls)) {
+            return Arrays.asList(CSMSettingsTag.ID, DefineTextFormatTag.ID);
         }
         if (TextTag.class.isAssignableFrom(cls)) {
             return Arrays.asList(CSMSettingsTag.ID);
@@ -1124,6 +1131,12 @@ public abstract class AbstractTagTree extends JTree {
         }
         if (cls == DefineButton2Tag.class) {
             return Arrays.asList(DefineButtonSoundTag.ID, DefineScalingGridTag.ID);
+        }
+        if (ImageTag.class.isAssignableFrom(cls)) {
+            return Arrays.asList(NameCharacterTag.ID);
+        }
+        if (cls == DefineSoundTag.class) {
+            return Arrays.asList(NameCharacterTag.ID);
         }
         return new ArrayList<>();
     }